Fix `@lsp` and diagnostic highlights
- Highlight `@lsp.mod.GlobalScope`
- Don't highlight `@lsp.type.macro`.
- Don't highlight `DiagnosticUnnecessary`, just underline it.
- Fix typo `@markup.strike` -> `@markup.strikethrough`

Highlighting `@lsp.type.macro` is annoying as hell when writing C.
Preprocessor macros are marked as global, so it's better to add a
highlight for `@lsp.mod.GlobalScope` instead.
Following the idea that parameters are bold, globals are now italized.
Many languages use all-caps as a stylistic convention for globals, so we
don't want to add a lot of highlighting on top.